GDBStub: Support watchpoints on linux
